+++ liquidctl/driver/nvidia.py
@@ -13,23 +13,105 @@ from liquidctl.util import RelaxedNamesEnum, check_unsafe
 
 _LOGGER = logging.getLogger(__name__)
 
+# sources for pci device and subsystem device ids:
+# - https://www.nv-drivers.eu/nvidia-all-devices.html
+# - https://pci-ids.ucw.cz/pci.ids
+# - https://gitlab.com/CalcProgrammer1/OpenRGB/-/blob/master/pci_ids/pci_ids.h
+
 # vendor, devices
 # FWUPD_GUID = [vendor]:[device] - use hwinfo to inspect
 NVIDIA = 0x10de
+NVIDIA_GTX_1050 = 0x1c81
+NVIDIA_GTX_1050_TI = 0x1c82
+NVIDIA_GTX_1060 = 0x1c03
 NVIDIA_GTX_1070 = 0x1b81
+NVIDIA_GTX_1070_TI = 0x1b82
 NVIDIA_GTX_1080 = 0x1b80
+NVIDIA_GTX_1080_TI = 0x1b06
+NVIDIA_GTX_1650 = 0x1f82
+NVIDIA_GTX_1650S = 0x2187
+NVIDIA_GTX_1660 = 0x2184
+NVIDIA_GTX_1660S = 0x21c4
+NVIDIA_GTX_1660_TI = 0x2182
+NVIDIA_RTX_2060S = 0x1f47
+NVIDIA_RTX_2060S_OC = 0x1f06
+NVIDIA_RTX_2060_TU104 = 0x1e89
+NVIDIA_RTX_2060_TU106 = 0x1f08
+NVIDIA_RTX_2070 = 0x1f02
+NVIDIA_RTX_2070S = 0x1ec7
+NVIDIA_RTX_2070S_OC = 0x1e84
+NVIDIA_RTX_2070_OC = 0x1f07
+NVIDIA_RTX_2080 = 0x1e82
+NVIDIA_RTX_2080S = 0x1e81
+NVIDIA_RTX_2080_REV_A = 0x1e87
+NVIDIA_RTX_2080_TI = 0x1e04
 NVIDIA_RTX_2080_TI_REV_A = 0x1e07
-# https://www.nv-drivers.eu/nvidia-all-devices.html
-# https://pci-ids.ucw.cz/pci.ids
+NVIDIA_RTX_3050 = 0x2507
+NVIDIA_RTX_3060 = 0x2503
+NVIDIA_RTX_3060_GA104 = 0x2487
+NVIDIA_RTX_3060_LHR = 0x2504
+NVIDIA_RTX_3060_TI = 0x2486
+NVIDIA_RTX_3060_TI_LHR = 0x2489
+NVIDIA_RTX_3070 = 0x2484
+NVIDIA_RTX_3070_LHR = 0x2488
+NVIDIA_RTX_3070_TI = 0x2482
+NVIDIA_RTX_3080 = 0x2206
+NVIDIA_RTX_3080_12G_LHR = 0x220a
+NVIDIA_RTX_3080_LHR = 0x2216
+NVIDIA_RTX_3080_TI = 0x2208
+NVIDIA_RTX_3090 = 0x2204
 
 # subsystem vendor ASUS, subsystem devices
 # PCI_SUBSYS_ID = [subsystem vendor]:[subsystem device] - use hwinfo to inspect
 ASUS = 0x1043
-ASUS_STRIX_GTX_1070 = 0x8599
+ASUS_STRIX_GTX_1050_O2G = 0x85d8
+ASUS_STRIX_GTX_1050_TI_O4G = 0x85cd
+ASUS_STRIX_GTX_1050_TI_O4G_2 = 0x85d1
+ASUS_STRIX_GTX_1060_6G = 0x85a4
+ASUS_STRIX_GTX_1060_O6G = 0x85ac
+ASUS_STRIX_GTX_1070 = 0x8598
+ASUS_STRIX_GTX_1070_OC = 0x8599
+ASUS_STRIX_GTX_1070_TI_8G = 0x861d
+ASUS_STRIX_GTX_1070_TI_A8G = 0x861e
+ASUS_STRIX_GTX_1080 = 0x8592
+ASUS_STRIX_GTX_1080_A8G = 0x85aa
+ASUS_STRIX_GTX_1080_O8G = 0x85f9
+ASUS_STRIX_GTX_1080_TI = 0x85eb
+ASUS_STRIX_GTX_1080_TI_11G = 0x85f1
+ASUS_STRIX_GTX_1080_TI_O11G = 0x85ea
+ASUS_STRIX_GTX_1080_TI_O11G_A02 = 0x85e4
+ASUS_STRIX_GTX_1650S_OC = 0x874f
+ASUS_STRIX_GTX_1660S_O6G = 0x8752
+ASUS_STRIX_GTX_1660_TI_OC = 0x86a5
+ASUS_STRIX_RTX_2060S_8G = 0x8730
+ASUS_STRIX_RTX_2060S_A8G = 0x86fc
+ASUS_STRIX_RTX_2060S_A8G_EVO = 0x8703
+ASUS_STRIX_RTX_2060S_O8G = 0x86fb
+ASUS_STRIX_RTX_2060_EVO = 0x86d3
+ASUS_STRIX_RTX_2060_O6G = 0x868e
+ASUS_STRIX_RTX_2060_O6G_EVO = 0x8775
+ASUS_STRIX_RTX_2070S_8G_8707 = 0x8707
+ASUS_STRIX_RTX_2070S_A8G = 0x8728
+ASUS_STRIX_RTX_2070S_A8G_86FF = 0x86ff
+ASUS_STRIX_RTX_2070S_A8G_8706 = 0x8706
+ASUS_STRIX_RTX_2070S_O8G = 0x8727
+ASUS_STRIX_RTX_2070S_O8G_8729 = 0x8729
+ASUS_STRIX_RTX_2070_A8G = 0x8671
+ASUS_STRIX_RTX_2070_O8G = 0x8670
+ASUS_STRIX_RTX_2080S_A8G = 0x8712
+ASUS_STRIX_RTX_2080S_O8G = 0x8711
+ASUS_STRIX_RTX_2080_O8G = 0x865f
+ASUS_STRIX_RTX_2080_TI_11G = 0x8687
 ASUS_STRIX_RTX_2080_TI_OC = 0x866a
+ASUS_TUF_RTX_3060_TI_O8G_OC = 0x87c6
 
 # subsystem vendor EVGA, subsystem devices
+# PCI_SUBSYS_ID = [subsystem vendor]:[subsystem device] - use hwinfo to inspect
 EVGA = 0x3842
+EVGA_GTX_1070_FTW = 0x6276
+EVGA_GTX_1070_FTW_DT_GAMING = 0x6274
+EVGA_GTX_1070_FTW_HYBRID = 0x6278
+EVGA_GTX_1070_TI_FTW2 = 0x6775
 EVGA_GTX_1080_FTW = 0x6286
 
 
@@ -92,6 +174,10 @@ class EvgaPascal(SmbusDriver, _NvidiaI2CDriver):
     _VENDOR = EVGA
     _ADDRESSES = [0x49]
     _MATCHES = [
+        (NVIDIA_GTX_1070, EVGA_GTX_1070_FTW, 'EVGA GTX 1070 FTW (experimental)'),
+        (NVIDIA_GTX_1070, EVGA_GTX_1070_FTW_DT_GAMING, 'EVGA GTX 1070 FTW DT Gaming (experimental)'),
+        (NVIDIA_GTX_1070, EVGA_GTX_1070_FTW_HYBRID, 'EVGA GTX 1070 FTW Hybrid (experimental)'),
+        (NVIDIA_GTX_1070_TI, EVGA_GTX_1070_TI_FTW2, 'EVGA GTX 1070 Ti FTW2 (experimental)'),
         (NVIDIA_GTX_1080, EVGA_GTX_1080_FTW, 'EVGA GTX 1080 FTW'),
     ]
 
@@ -115,8 +201,15 @@ class EvgaPascal(SmbusDriver, _NvidiaI2CDriver):
         for dev_id, sub_dev_id, desc in pre_probed:
             dev = cls(smbus, desc, vendor_id=EVGA, product_id=EVGA_GTX_1080_FTW,
                       address=cls._ADDRESSES[0])
-            _LOGGER.debug('instantiated %s driver for %s', cls.__name__, desc)
+            _LOGGER.debug('found %s: %s', cls.__name__, desc)
             yield dev
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if 'experimental' in self.description:
+            self._UNSAFE = ['smbus', 'experimental_evga_gpu']
+        else:
+            self._UNSAFE = ['smbus']
 
     def get_status(self, verbose=False, **kwargs):
         """Get a status report.
@@ -130,9 +223,9 @@ class EvgaPascal(SmbusDriver, _NvidiaI2CDriver):
         if not verbose:
             return []
 
-        if not check_unsafe('smbus', **kwargs):
-            _LOGGER.warning("%s: nothing returned, requires unsafe feature 'smbus'",
-                            self.description)
+        if not check_unsafe(*self._UNSAFE, **kwargs):
+            _LOGGER.warning("%s: disabled, requires unsafe features '%s'",
+                            self.description, ','.join(self._UNSAFE))
             return []
 
         mode = self.Mode(self._smbus.read_byte_data(self._address, self._REG_MODE))
@@ -169,7 +262,7 @@ class EvgaPascal(SmbusDriver, _NvidiaI2CDriver):
         preferable, if the use case allows for them.
         """
 
-        check_unsafe('smbus', error=True, **kwargs)
+        check_unsafe(*self._UNSAFE, error=True, **kwargs)
 
         colors = list(colors)
 
@@ -225,10 +318,57 @@ class RogTuring(SmbusDriver, _NvidiaI2CDriver):
     _VENDOR = ASUS
     _ADDRESSES = [0x29, 0x2a, 0x60]
     _MATCHES = [
-        (NVIDIA_GTX_1070, ASUS_STRIX_GTX_1070,
-            'ASUS Strix GTX 1070'),
-        (NVIDIA_RTX_2080_TI_REV_A, ASUS_STRIX_RTX_2080_TI_OC,
-            'ASUS Strix RTX 2080 Ti OC'),
+        # description normalization rules:
+        # - remove redundant ROG and meaningless GAMING;
+        # - remove redundancies within a particular GPU core;
+        # - keep OC when appropriate;
+        # - uppercase ASUS and TUF, titlecase Evo, Ti and Stix;
+        # - enforce ASUS Strix|TUF GTX|RTX <number> [Super] [Ti] [Evo] [OC] [...] order;
+        # - use consistent memory sizes: 2GB, 4GB, 6GB, 8GB (but omit when redundant).
+
+        # stable
+        (NVIDIA_GTX_1070, ASUS_STRIX_GTX_1070_OC, 'ASUS Strix GTX 1070 OC'),
+        (NVIDIA_RTX_2080_TI_REV_A, ASUS_STRIX_RTX_2080_TI_OC, 'ASUS Strix RTX 2080 Ti OC'),
+
+        # experimental
+        (NVIDIA_GTX_1050, ASUS_STRIX_GTX_1050_O2G, 'ASUS Strix GTX 1050 OC (experimental)'),
+        (NVIDIA_GTX_1050_TI, ASUS_STRIX_GTX_1050_TI_O4G, 'ASUS Strix GTX 1050 Ti OC (experimental)'),
+        (NVIDIA_GTX_1050_TI, ASUS_STRIX_GTX_1050_TI_O4G_2, 'ASUS Strix GTX 1050 Ti OC (experimental)'),
+        (NVIDIA_GTX_1060, ASUS_STRIX_GTX_1060_6G, 'ASUS Strix GTX 1060 6GB (experimental)'),
+        (NVIDIA_GTX_1060, ASUS_STRIX_GTX_1060_O6G, 'ASUS Strix GTX 1060 OC 6GB (experimental)'),
+        (NVIDIA_GTX_1070, ASUS_STRIX_GTX_1070, 'ASUS Strix GTX 1070 (experimental)'),
+        (NVIDIA_GTX_1070_TI, ASUS_STRIX_GTX_1070_TI_8G, 'ASUS Strix GTX 1070 Ti (experimental)'),
+        (NVIDIA_GTX_1070_TI, ASUS_STRIX_GTX_1070_TI_A8G, 'ASUS Strix GTX 1070 Ti Advanced (experimental)'),
+        (NVIDIA_GTX_1080, ASUS_STRIX_GTX_1080, 'ASUS Strix GTX 1080 (experimental)'),
+        (NVIDIA_GTX_1080, ASUS_STRIX_GTX_1080_A8G, 'ASUS Strix GTX 1080 Advanced (experimental)'),
+        (NVIDIA_GTX_1080, ASUS_STRIX_GTX_1080_O8G, 'ASUS Strix GTX 1080 OC (experimental)'),
+        (NVIDIA_GTX_1080_TI, ASUS_STRIX_GTX_1080_TI, 'ASUS Strix GTX 1080 Ti (experimental)'),
+        (NVIDIA_GTX_1080_TI, ASUS_STRIX_GTX_1080_TI_11G, 'ASUS Strix GTX 1080 Ti (experimental)'),
+        (NVIDIA_GTX_1080_TI, ASUS_STRIX_GTX_1080_TI_O11G, 'ASUS Strix GTX 1080 Ti OC (experimental)'),
+        (NVIDIA_GTX_1080_TI, ASUS_STRIX_GTX_1080_TI_O11G_A02, 'ASUS Strix GTX 1080 Ti OC (experimental)'),
+        (NVIDIA_GTX_1650S, ASUS_STRIX_GTX_1650S_OC, 'ASUS Strix GTX 1650 Super OC (experimental)'),
+        (NVIDIA_GTX_1660S, ASUS_STRIX_GTX_1660S_O6G, 'ASUS Strix GTX 1660 Super OC (experimental)'),
+        (NVIDIA_GTX_1660_TI, ASUS_STRIX_GTX_1660_TI_OC, 'ASUS Strix GTX 1660 Ti OC (experimental)'),
+        (NVIDIA_RTX_2060S, ASUS_STRIX_RTX_2060S_A8G_EVO, 'ASUS Strix RTX 2060 Super Evo Advanced (experimental)'),
+        (NVIDIA_RTX_2060S_OC, ASUS_STRIX_RTX_2060S_8G, 'ASUS Strix RTX 2060 Super (experimental)'),
+        (NVIDIA_RTX_2060S_OC, ASUS_STRIX_RTX_2060S_A8G, 'ASUS Strix RTX 2060 Super Advanced (experimental)'),
+        (NVIDIA_RTX_2060S_OC, ASUS_STRIX_RTX_2060S_O8G, 'ASUS Strix RTX 2060 Super OC (experimental)'),
+        (NVIDIA_RTX_2060_TU104, ASUS_STRIX_RTX_2060_O6G_EVO, 'ASUS Strix RTX 2060 Evo OC (experimental)'),
+        (NVIDIA_RTX_2060_TU106, ASUS_STRIX_RTX_2060_EVO, 'ASUS Strix RTX 2060 Evo (experimental)'),
+        (NVIDIA_RTX_2060_TU106, ASUS_STRIX_RTX_2060_O6G, 'ASUS Strix RTX 2060 OC (experimental)'),
+        (NVIDIA_RTX_2070S, ASUS_STRIX_RTX_2070S_A8G_86FF, 'ASUS Strix RTX 2070 Super Advanced (experimental)'),
+        (NVIDIA_RTX_2070S_OC, ASUS_STRIX_RTX_2070S_8G_8707, 'ASUS Strix RTX 2070 (experimental)'),
+        (NVIDIA_RTX_2070S_OC, ASUS_STRIX_RTX_2070S_A8G, 'ASUS Strix RTX 2070 Super Advanced (experimental)'),
+        (NVIDIA_RTX_2070S_OC, ASUS_STRIX_RTX_2070S_A8G_8706, 'ASUS Strix RTX 2070 Super Advanced (experimental)'),
+        (NVIDIA_RTX_2070S_OC, ASUS_STRIX_RTX_2070S_O8G, 'ASUS Strix RTX 2070 Super OC (experimental)'),
+        (NVIDIA_RTX_2070S_OC, ASUS_STRIX_RTX_2070S_O8G_8729, 'ASUS Strix RTX 2070 Super OC (experimental)'),
+        (NVIDIA_RTX_2070_OC, ASUS_STRIX_RTX_2070_A8G, 'ASUS Strix RTX 2070 Advanced (experimental)'),
+        (NVIDIA_RTX_2070_OC, ASUS_STRIX_RTX_2070_O8G, 'ASUS Strix RTX 2070 OC (experimental)'),
+        (NVIDIA_RTX_2080S, ASUS_STRIX_RTX_2080S_A8G, 'ASUS Strix RTX 2080 Super Advanced (experimental)'),
+        (NVIDIA_RTX_2080S, ASUS_STRIX_RTX_2080S_O8G, 'ASUS Strix RTX 2080 Super OC (experimental)'),
+        (NVIDIA_RTX_2080_REV_A, ASUS_STRIX_RTX_2080_O8G, 'ASUS Strix RTX 2080 OC (experimental)'),
+        (NVIDIA_RTX_2080_TI, ASUS_STRIX_RTX_2080_TI_11G, 'ASUS Strix RTX 2080 Ti (experimental)'),
+        (NVIDIA_RTX_3060_TI_LHR, ASUS_TUF_RTX_3060_TI_O8G_OC, 'ASUS TUF RTX 3060 Ti OC (experimental)'),
     ]
 
     _SENTINEL_ADDRESS = 0xffff  # intentionally invalid
@@ -281,6 +421,13 @@ class RogTuring(SmbusDriver, _NvidiaI2CDriver):
                               cls.__name__, desc, selected_address)
                 yield dev
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if 'experimental' in self.description:
+            self._UNSAFE = ['smbus', 'experimental_asus_gpu']
+        else:
+            self._UNSAFE = ['smbus']
+
     def get_status(self, verbose=False, **kwargs):
         """Get a status report.
 
@@ -293,9 +440,9 @@ class RogTuring(SmbusDriver, _NvidiaI2CDriver):
         if not verbose:
             return []
 
-        if not check_unsafe('smbus', **kwargs):
-            _LOGGER.warning("%s: nothing returned, requires unsafe feature 'smbus'",
-                            self.description)
+        if not check_unsafe(*self._UNSAFE, **kwargs):
+            _LOGGER.warning("%s: disabled, requires unsafe features '%s'",
+                            self.description, ','.join(self._UNSAFE))
             return []
 
         assert self._address != self._SENTINEL_ADDRESS, \
@@ -343,7 +490,7 @@ class RogTuring(SmbusDriver, _NvidiaI2CDriver):
 
         """
 
-        check_unsafe('smbus', error=True, **kwargs)
+        check_unsafe(*self._UNSAFE, error=True, **kwargs)
 
         assert self._address != self._SENTINEL_ADDRESS, \
                'invalid address (probing may not have had access to SMbus)'