+++ tests/test_smart_device.py
@@ -1,6 +1,7 @@
 import pytest
 from _testutils import MockHidapiDevice, Report
 
+from liquidctl.driver.hwmon import HwmonDevice
 from liquidctl.driver.smart_device import SmartDevice
 
 SAMPLE_RESPONSES = [
@@ -31,7 +32,7 @@ def test_smart_device_constructor(mockSmartDevice):
 def test_smart_device_not_totally_broken(mockSmartDevice):
     dev = mockSmartDevice
 
-    for i in range(3):
+    for i in range(4):
         dev.device.preload_read(Report(0, bytes(63)))
 
     dev.initialize()
@@ -43,8 +44,39 @@ def test_smart_device_not_totally_broken(mockSmartDevice):
     dev.set_fixed_speed(channel='fan3', duty=50)
 
 
-def test_smart_device_reads_status(mockSmartDevice):
+@pytest.mark.parametrize('has_hwmon,direct_access', [(False, False), (True, True), (True, False)])
+def test_smart_device_initializes(mockSmartDevice, has_hwmon, direct_access, tmp_path):
     dev = mockSmartDevice
+    if has_hwmon:
+        dev._hwmon = HwmonDevice('mock_module', tmp_path)
+
+    for _, capdata in enumerate(SAMPLE_RESPONSES):
+        capdata = bytes.fromhex(capdata)
+        dev.device.preload_read(Report(capdata[0], capdata[1:]))
+
+    expected = [
+        ('Firmware version', '1.0.7', ''),
+        ('LED accessories', 2, ''),
+        ('LED accessory type', 'HUE+ Strip', ''),
+        ('LED count (total)', 20, ''),
+    ]
+
+    got = dev.initialize(direct_access=direct_access)
+
+    assert expected == got
+
+    writes = len(dev.device.sent)
+    if not has_hwmon or direct_access:
+        assert writes == 2
+    else:
+        assert writes == 0
+
+
+@pytest.mark.parametrize('has_hwmon,direct_access', [(False, False), (True, True)])
+def test_smart_device_reads_status_directly(mockSmartDevice, has_hwmon, direct_access):
+    dev = mockSmartDevice
+    if has_hwmon:
+        dev._hwmon = HwmonDevice(None, None)
 
     for _, capdata in enumerate(SAMPLE_RESPONSES):
         capdata = bytes.fromhex(capdata)
@@ -66,11 +98,47 @@ def test_smart_device_reads_status(mockSmartDevice):
         ('Fan 3 voltage', 11.91, 'V'),
         ('Fan 3 current', 0.03, 'A'),
         ('Fan 3 control mode', None, ''),
-        ('Firmware version', '1.0.7', ''),
-        ('LED accessories', 2, ''),
-        ('LED accessory type', 'HUE+ Strip', ''),
-        ('LED count (total)', 20, ''),
         ('Noise level', 63, 'dB')
+    ]
+
+    got = dev.get_status(direct_access=direct_access)
+
+    assert expected == got
+
+
+def test_smart_device_reads_status_from_hwmon(mockSmartDevice, tmp_path):
+    dev = mockSmartDevice
+
+    dev._hwmon = HwmonDevice('mock_module', tmp_path)
+    (tmp_path / 'fan1_input').write_text('1461\n')
+    (tmp_path / 'in0_input').write_text('11910\n')
+    (tmp_path / 'curr1_input').write_text('20\n')
+    (tmp_path / 'pwm1_mode').write_text('1\n')
+    (tmp_path / 'fan2_input').write_text('1336\n')
+    (tmp_path / 'in1_input').write_text('11910\n')
+    (tmp_path / 'curr2_input').write_text('20\n')
+    (tmp_path / 'pwm2_mode').write_text('0\n')
+    (tmp_path / 'fan3_input').write_text('1390\n')
+    (tmp_path / 'in2_input').write_text('11910\n')
+    (tmp_path / 'curr3_input').write_text('30\n')
+    (tmp_path / 'pwm3_mode').write_text('1\n')
+
+    # skip initialize for now, we're not emulating the behavior precisely
+    # enough to require it here
+
+    expected = [
+        ('Fan 1 speed', 1461, 'rpm'),
+        ('Fan 1 voltage', 11.91, 'V'),
+        ('Fan 1 current', 0.02, 'A'),
+        ('Fan 1 control mode', 'PWM', ''),
+        ('Fan 2 speed', 1336, 'rpm'),
+        ('Fan 2 voltage', 11.91, 'V'),
+        ('Fan 2 current', 0.02, 'A'),
+        ('Fan 2 control mode', 'DC', ''),
+        ('Fan 3 speed', 1390, 'rpm'),
+        ('Fan 3 voltage', 11.91, 'V'),
+        ('Fan 3 current', 0.03, 'A'),
+        ('Fan 3 control mode', 'PWM', ''),
     ]
 
     got = dev.get_status()